add cast to kill warning in previous commit
git-svn-id: svn://localhost/svn/u-boot/trunk@1454 e8ad1c53-384c-47db-b5ce-66ae1f620dea

+++ u-boot-2008.10/common/cmd_elf.c
@@ -296,7 +296,7 @@ unsigned long load_elf_image (unsigned long addr)
 		printf ("Loading phdr %i to 0x%p (%i bytes)\n",
 		        i, dst, phdr->p_filesz);
 		memcpy (dst, src, phdr->p_filesz);
-		flush_cache (dst, phdr->p_filesz);
+		flush_cache ((unsigned long)dst, phdr->p_filesz);
 		++phdr;
 	}
